feat(bit-reversal): change int to size_t

+++ src/utils/bit-reversal.cpp
@@ -2,10 +2,10 @@
 
 std::vector<std::complex<double>> & sequentialBitReversal(std::vector<std::complex<double>> &data) {
     const size_t N = data.size();
-    const int log_N = static_cast<int>(std::log2(N));
+    const size_t log_N = std::log2(N);
     for (size_t i = 0; i < N; ++i) {
         size_t reversed_index = 0;
-        for (int j = 0; j < log_N; ++j) {
+        for (size_t j = 0; j < log_N; ++j) {
             if (i & (1 << j)) {
                 reversed_index |= (1 << (log_N - j - 1));
             }
@@ -19,11 +19,11 @@ std::vector<std::complex<double>> & sequentialBitReversal(std::vector<std::compl
 
 std::vector<std::complex<double>> & parallelBitReversal(std::vector<std::complex<double>> &data) {
     const size_t N = data.size();
-    const int log_N = static_cast<int>(std::log2(N));
+    const size_t log_N = std::log2(N);
 #pragma omp parallel for
-    for (int i = 0; i < N; ++i) {
-        int reversed_index = 0;
-        for (int j = 0; j < log_N; ++j) {
+    for (size_t i = 0; i < N; ++i) {
+        size_t reversed_index = 0;
+        for (size_t j = 0; j < log_N; ++j) {
             if (i & (1 << j)) {
                 reversed_index |= (1 << (log_N - j - 1));
             }
